Refactored Tratis/ThanksListWidget
Combined the 2 lists into a single widget, as having both was redundent.
In the process, fixed issue #95
Also, fixed a bug with a use of `DebugPrint` that prevented reseting the data.
